reorganize org files, edit tests
