+++ app/pages/organization/organization-page.jsx
@@ -3,121 +3,12 @@ import classnames from 'classnames';
 import { Markdown } from 'markdownz';
 import { Link } from 'react-router';
 import Translate from 'react-translate-component';
+import OrganizationProjectCards from './organization-project-cards';
+import OrganizationMetadata from './organization-metadata';
 import Thumbnail from '../../components/thumbnail';
 import SocialIcons from '../../lib/social-icons';
-import ProjectCard from '../../partials/project-card';
 
 const AVATAR_SIZE = 100;
-
-export const OrganizationProjectCards = ({ errorFetchingProjects, fetchingProjects, projects, projectAvatars }) => {
-  if (fetchingProjects) {
-    return (
-      <div className="organization-page__projects-status">
-        <p><Translate content="organization.home.projects.loading" /></p>
-      </div>);
-  } else if (errorFetchingProjects) {
-    return (
-      <div className="organization-page__projects-status">
-        <p><Translate content="organization.home.projects.error" /></p>
-        <p>
-          <code>{errorFetchingProjects.toString()}</code>
-        </p>
-      </div>);
-  } else if (!fetchingProjects && !projects.length) {
-    return (
-      <div className="organization-page__projects-status">
-        <p><Translate content="organization.home.projects.none" /></p>
-      </div>);
-  } else {
-    return (
-      <div className="project-card-list">
-        {projects.map((project) => {
-          let projectAvatar = projectAvatars.find(avatar => avatar.links.linked.id === project.id);
-          if (!projectAvatar) {
-            projectAvatar = { src: '' };
-          }
-          return (
-            <ProjectCard key={project.id} project={project} imageSrc={projectAvatar.src} />);
-        })}
-      </div>);
-  }
-};
-
-OrganizationProjectCards.propTypes = {
-  errorFetchingProjects: React.PropTypes.shape({
-    message: React.PropTypes.string
-  }),
-  fetchingProjects: React.PropTypes.bool,
-  projects: React.PropTypes.arrayOf(
-    React.PropTypes.shape({
-      id: React.PropTypes.string,
-      display_name: React.PropTypes.string
-    })
-  ),
-  projectAvatars: React.PropTypes.arrayOf(
-    React.PropTypes.shape({
-      id: React.PropTypes.string,
-      src: React.PropTypes.string
-    })
-  )
-};
-
-export const OrganizationMetaData = ({ displayName, projects }) => {
-  function extractStat(statName) {
-    return projects.reduce((accum, project) => {
-      if (project[statName]) {
-        return accum + project[statName];
-      } else {
-        return accum;
-      }
-    }, 0);
-  }
-
-  return (
-    <div className="organization-page__container">
-      <div className="project-metadata">
-        <span className="organization-details__heading">
-          {displayName}{' '}<Translate content="project.home.metadata.statistics" />
-        </span>
-        <div className="project-metadata-stats">
-          <div className="project-metadata-stat">
-            <div className="project-metadata-stat__value">{projects.length.toLocaleString()}</div>
-            <div className="project-metadata-stat__label">
-              <Translate content="organization.home.metadata.projects" />
-            </div>
-          </div>
-          <div className="project-metadata-stat">
-            <div className="project-metadata-stat__value">{extractStat('subjects_count').toLocaleString()}</div>
-            <div className="project-metadata-stat__label">
-              <Translate content="project.home.metadata.subjects" />
-            </div>
-          </div>
-          <div className="project-metadata-stat">
-            <div className="project-metadata-stat__value">{extractStat('classifications_count').toLocaleString()}</div>
-            <div className="project-metadata-stat__label">
-              <Translate content="project.home.metadata.classifications" />
-            </div>
-          </div>
-          <div className="project-metadata-stat">
-            <div className="project-metadata-stat__value">{extractStat('retired_subjects_count').toLocaleString()}</div>
-            <div className="project-metadata-stat__label">
-              <Translate content="project.home.metadata.completedSubjects" />
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>);
-};
-
-OrganizationMetaData.propTypes = {
-  displayName: React.PropTypes.string,
-  projects: React.PropTypes.arrayOf(
-    React.PropTypes.shape({
-      id: React.PropTypes.string,
-      display_name: React.PropTypes.string
-    })
-  )
-};
 
 class OrganizationPage extends React.Component {
   constructor() {
@@ -253,7 +144,7 @@ class OrganizationPage extends React.Component {
             </div>
           </div>
 
-          <OrganizationMetaData
+          <OrganizationMetadata
             displayName={this.props.organization.display_name}
             projects={this.props.organizationProjects}
           />
