+++ app/pages/organization/organization-page.spec.js
@@ -5,14 +5,8 @@ import React from 'react';
 import assert from 'assert';
 import { shallow } from 'enzyme';
 import sinon from 'sinon';
-import Translate from 'react-translate-component';
-import OrganizationPage, { OrganizationProjectCards, OrganizationMetaData } from './organization-page';
+import OrganizationPage from './organization-page';
 import { organization } from './organization-container.spec';
-
-const organizationProjects = [1, 2, 3].map(i => ({
-  id: i.toString(),
-  display_name: `Project #${i}`
-}));
 
 const organizationPages = [{
   content: 'test content',
@@ -86,47 +80,6 @@ describe('OrganizationPage', function () {
     assert.equal(cards.length, 1);
   });
 
-  it('should show loading projects message if fetchingProjects', function () {
-    const message = <Translate content="organization.home.projects.loading" />;
-    const wrapper = shallow(
-      <OrganizationProjectCards fetchingProjects={true} />);
-    const status = wrapper.find('.organization-page__projects-status');
-    assert.equal(status.length, 1);
-    assert.equal(status.contains(message), true);
-  });
-
-  it('should show projects error message if errorFetchingProjects', function () {
-    const message = <Translate content="organization.home.projects.error" />;
-    const wrapper = shallow(
-      <OrganizationProjectCards errorFetchingProjects={{ message: 'test error' }} />);
-    const status = wrapper.find('.organization-page__projects-status');
-    assert.equal(status.length, 1);
-    assert.equal(status.contains(message), true);
-  });
-
-  it('should show no projects associated message if fetchingProjects false and no projects', function () {
-    const message = <Translate content="organization.home.projects.none" />;
-    const wrapper = shallow(
-      <OrganizationProjectCards
-        fetchingProjects={false}
-        projects={[]}
-      />);
-    const status = wrapper.find('.organization-page__projects-status');
-    assert.equal(status.length, 1);
-    assert.equal(status.contains(message), true);
-  });
-
-  it('should show project cards once fetched', function () {
-    const wrapper = shallow(
-      <OrganizationProjectCards
-        fetchingProjects={false}
-        projects={organizationProjects}
-        projectAvatars={[]}
-      />);
-    const cards = wrapper.find('ProjectCard');
-    assert.equal(cards.length, organizationProjects.length);
-  });
-
   it('should show category buttons if the organization has categories', function () {
     const wrapper = shallow(<OrganizationPage organization={organization} />);
     const categories = wrapper.find('.organization-page__category-button');
@@ -151,15 +104,6 @@ describe('OrganizationPage', function () {
     const wrapper = shallow(<OrganizationPage organization={organization} />);
     const quote = wrapper.find('.organization-researcher-words');
     assert.equal(quote.length, 0);
-
-  it('should render a stats section', function () {
-    const wrapper = shallow(
-      <OrganizationMetaData
-        displayName={organization.display_name}
-        projects={organizationProjects}
-      />);
-    const links = wrapper.find('.project-metadata-stat');
-    assert.equal(links.length, 4);
   });
 
   describe('with pages about content', function () {
